# test/UnitTest/OpenApi/OpenApiTests.VerifyFullSwagger.verified.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Altinn Platform Storage - complete",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://platform.tt02.altinn.no",
      "description": "TT02"
    },
    {
      "url": "https://platform.altinn.no",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/storage/api/v1/instances/{instanceOwnerPartyId}/{instanceGuid}": {
      "get": {
        "tags": [
          "Instances"
        ],
        "summary": "Gets a specific instance with the given instance id.\nThis endpoint is maintained for backwards compatibility but requires an unnecessary instanceOwnerPartyId parameter.\nPrefer using GetByGuid which only requires the instance GUID.",
        "parameters": [
          {
            "name": "instanceOwnerPartyId",
            "in": "path",
            "description": "The party id of the instance owner.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "instanceGuid",
            "in": "path",
            "description": "The id of the instance to retrieve.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Instance"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Instances"
        ],
        "summary": "Delete an instance.",
        "parameters": [
          {
            "name": "instanceOwnerPartyId",
            "in": "path",
            "description": "The party id of the instance owner.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "instanceGuid",
            "in": "path",
            "description": "The id of the instance that should be deleted.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "hard",
            "in": "query",
            "description": "if true hard delete will take place. if false, the instance gets its status.softDelete attribute set to current date and time.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "If-Instance-Version-Match",
            "in": "header",
            "description": "Optional expected aggregate instance version.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Process-State-Version-Match",
            "in": "header",
            "description": "Optional expected process-state version.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Instance"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/storage/api/v1/instances/{instanceGuid}": {
      "get": {
        "tags": [
          "Instances"
        ],
        "summary": "Gets a specific instance with the given instance guid.\nPrefer this endpoint over Get which requires an unnecessary instanceOwnerPartyId parameter.",
        "parameters": [
          {
            "name": "instanceGuid",
            "in": "path",
            "description": "The id of the instance to retrieve.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Instance"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/storage/api/v1/sbl/instances/{instanceOwnerPartyId}/{instanceGuid}": {
      "get": {
        "tags": [
          "MessageBoxInstances"
        ],
        "summary": "Gets all instances in a given state for a given instance owner.",
        "parameters": [
          {
            "name": "instanceOwnerPartyId",
            "in": "path",
            "description": "the instance owner id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "instanceGuid",
            "in": "path",
            "description": "the instance guid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "language id en, nb, nn-NO\"",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "MessageBoxInstances"
        ],
        "summary": "Marks an instance for deletion in storage.",
        "parameters": [
          {
            "name": "instanceGuid",
            "in": "path",
            "description": "instance id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "instanceOwnerPartyId",
            "in": "path",
            "description": "instance owner",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "hard",
            "in": "query",
            "description": "if true is marked for hard delete.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/storage/api/v1/studio/instances/{org}/{app}": {
      "get": {
        "tags": [
          "StudioInstances"
        ],
        "summary": "Retrieves all (simplified) instances that match the specified query parameters. Parameters can be combined. Invalid or unknown parameter values will result in a 400 Bad Request response.",
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "description": "The organization identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "The application identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archiveReference",
            "in": "query",
            "description": "Gets or sets the archive reference.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "process.currentTask",
            "in": "query",
            "description": "The current task identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "process.isComplete",
            "in": "query",
            "description": "A value indicating whether the process is completed.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "lastChanged",
            "in": "query",
            "description": "The last changed date.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "created",
            "in": "query",
            "description": "The creation date.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "confirmed",
            "in": "query",
            "description": "Confirmed = false is a compact version of ExcludeConfirmedBy indicating\nExcludeConfirmedBy for the org that invokes the request",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "status.isSoftDeleted",
            "in": "query",
            "description": "A value indicating whether the instance is soft deleted.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "status.isHardDeleted",
            "in": "query",
            "description": "A value indicating whether the instance is hard deleted.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "status.isArchived",
            "in": "query",
            "description": "A value indicating whether the instance is archived.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "continuationToken",
            "in": "query",
            "description": "The continuation token.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleInstanceQueryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/storage/api/v1/studio/instances/{org}/{app}/{instanceGuid}": {
      "get": {
        "tags": [
          "StudioInstances"
        ],
        "summary": "Gets a specific instance with the given instance id.",
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "description": "The org owning the the instance to retrieve.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "The app tied to the instance to retrieve.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceGuid",
            "in": "path",
            "description": "The id of the instance to retrieve.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleInstanceDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StudioInstances"
        ],
        "summary": "Deletes an instance with the given instance id. Soft deletion.",
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "description": "The org owning the instance.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app",
            "in": "path",
            "description": "The app tied to the instance.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceGuid",
            "in": "path",
            "description": "The id of the instance.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Instance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "instanceOwner": {
            "$ref": "#/components/schemas/InstanceOwner"
          },
          "appId": {
            "type": "string",
            "nullable": true
          },
          "org": {
            "type": "string",
            "nullable": true
          },
          "selfLinks": {
            "$ref": "#/components/schemas/ResourceLinks"
          },
          "dueBefore": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "visibleAfter": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "process": {
            "$ref": "#/components/schemas/ProcessState"
          },
          "status": {
            "$ref": "#/components/schemas/InstanceStatus"
          },
          "completeConfirmations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompleteConfirmation"
            },
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataElement"
            },
            "nullable": true
          },
          "presentationTexts": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          },
          "dataValues": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "lastChanged": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastChangedBy": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "InstanceOwner": {
        "type": "object",
        "properties": {
          "partyId": {
            "type": "string",
            "nullable": true
          },
          "personNumber": {
            "type": "string",
            "nullable": true
          },
          "organisationNumber": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "externalIdentifier": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "ResourceLinks": {
        "type": "object",
        "properties": {
          "apps": {
            "type": "string",
            "nullable": true
          },
          "platform": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SimpleInstanceDetails": {
        "required": [
          "app",
          "id",
          "org"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SimpleDataElement"
            },
            "description": "A list of simplified data elements associated with the instance",
            "nullable": true
          },
          "id": {
            "type": "string",
            "description": "The unique id of the instance {instanceGuid}.",
            "nullable": true
          },
          "org": {
            "type": "string",
            "description": "Application owner identifier, usually a abbreviation of organisation name. All in lower case.",
            "nullable": true
          },
          "app": {
            "type": "string",
            "description": "The id of the application this is an instance of.",
            "nullable": true
          },
          "isRead": {
            "type": "boolean",
            "description": "If the instance is read."
          },
          "currentTaskId": {
            "type": "string",
            "description": "Reference to the current task/event element id as given in the process definition, Altinn.Platform.Storage.Interface.Models.ProcessElementInfo.",
            "nullable": true
          },
          "currentTaskName": {
            "type": "string",
            "description": "The name of the process element, Altinn.Platform.Storage.Interface.Models.ProcessElementInfo.",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "description": "The date the instance process was completed.",
            "format": "date-time",
            "nullable": true
          },
          "archivedAt": {
            "type": "string",
            "description": "The date the instance was archived.",
            "format": "date-time",
            "nullable": true
          },
          "softDeletedAt": {
            "type": "string",
            "description": "The date the instance was deleted.",
            "format": "date-time",
            "nullable": true
          },
          "hardDeletedAt": {
            "type": "string",
            "description": "The date the instance was marked for hard delete by user.",
            "format": "date-time",
            "nullable": true
          },
          "confirmedAt": {
            "type": "string",
            "description": "The date the instance was marked as confirmed by service owner.",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "The date and time for when the instance was created.",
            "format": "date-time",
            "nullable": true
          },
          "lastChangedAt": {
            "type": "string",
            "description": "The date and time for when the instance was last edited.",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents a simplified instance with most fields redacted, see Altinn.Platform.Storage.Interface.Models.Instance."
      },
      "SimpleInstanceQueryResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "description": "The number of items in this response.",
            "format": "int64"
          },
          "self": {
            "type": "string",
            "description": "The current query.",
            "nullable": true
          },
          "next": {
            "type": "string",
            "description": "A link to the next page.",
            "nullable": true
          },
          "instances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SimpleInstance"
            },
            "description": "The metadata.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Query response object"
      },
...
    },
    "securitySchemes": {
      "JwtCookie": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\". Remember to add \"Bearer\" to the input below before your token.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "JwtCookie": []
    }
  ],
...
}